refactor: decouple the flex_catalog core models and update its definition
- add the flex_catalog app for testing and quality checks
- update the setup.py file to include the flex_catalog logic
