+++ corporate_partner_access/models.py
@@ -1,10 +1,10 @@
 """Models for managing course catalogs and access for corporate partners."""
 
-
 from django.contrib.auth import get_user_model
 from django.db import models
 
 from corporate_partner_access.edxapp_wrapper.course_module import course_overview
+from flex_catalog.models import FlexibleCatalogModel
 
 
 class CorporatePartner(models.Model):
@@ -17,12 +17,64 @@ class CorporatePartner(models.Model):
     homepage_url = models.URLField(blank=True, null=True)
 
     class Meta:
+        """Meta options for CorporatePartner model."""
+
         verbose_name = "Corporate Partner"
         verbose_name_plural = "Corporate Partners"
         ordering = ["name"]
 
     def __str__(self):
+        """Return a string representation of the CorporatePartner instance."""
         return f"<CorporatePartner: {self.name} (Code: {self.code})>"
+
+
+class CorporatePartnerCatalog(FlexibleCatalogModel):
+    """Catalog model for corporate partners."""
+
+    corporate_partner = models.ForeignKey(
+        "CorporatePartner",
+        on_delete=models.CASCADE,
+        related_name="catalogs",
+    )
+    course_enrollment_limit = models.PositiveIntegerField(
+        default=0, help_text="Max enrollments allowed in this catalog."
+    )
+    user_limit = models.PositiveIntegerField(
+        default=0, help_text="Max users allowed in this catalog."
+    )
+    is_self_enrollment = models.BooleanField(default=False)
+    available_start_date = models.DateTimeField(null=True, blank=True)
+    available_end_date = models.DateTimeField(null=True, blank=True)
+    custom_courses = models.BooleanField(
+        default=False, help_text="If True, allows custom courses."
+    )
+    authorization_additional_message = models.TextField(blank=True, null=True)
+    support_email = models.EmailField(blank=True, null=True)
+    is_public = models.BooleanField(default=False)
+    catalog_alternative_link = models.URLField(blank=True, null=True)
+
+    courses = models.ManyToManyField(
+        course_overview(),
+        through="CorporatePartnerCatalogCourse",
+        related_name="partner_catalogs",
+    )
+
+    learners = models.ManyToManyField(
+        get_user_model(),
+        through="CorporatePartnerCatalogLearner",
+        related_name="enrolled_catalogs",
+    )
+
+    class Meta:
+        """Meta options for CorporatePartnerCatalog model."""
+
+        verbose_name = "Corporate Partner Catalog"
+        verbose_name_plural = "Corporate Partner Catalogs"
+        ordering = ["name"]
+
+    def get_courses(self):
+        """Return all catalog courses associated with this instance."""
+        return self.courses.all()
 
 
 class CorporatePartnerCatalogEmailRegex(models.Model):
@@ -39,6 +91,7 @@ class CorporatePartnerCatalogEmailRegex(models.Model):
     )
 
     def __str__(self):
+        """Return a string representation of the CorporatePartnerCatalogEmailRegex instance."""
         return (
             f"<CorporatePartnerCatalogEmailRegex: {self.catalog.name} - {self.regex}>"
         )
@@ -55,12 +108,15 @@ class CorporatePartnerCatalogCourse(models.Model):
     )
 
     class Meta:
+        """Meta options for CorporatePartnerCatalogCourse model."""
+
         verbose_name = "Corporate Partner Catalog Course"
         verbose_name_plural = "Corporate Partner Catalog Courses"
         unique_together = ("catalog", "course_overview")
         ordering = ["position"]
 
     def __str__(self):
+        """Return string representation of the CorporatePartnerCatalogCourse instance."""
         return f"<CorporatePartnerCatalogCourse: {self.course_overview.display_name}>"  # pylint: disable=no-member
 
 
@@ -73,12 +129,13 @@ class CorporatePartnerCatalogLearner(models.Model):
     active = models.BooleanField(default=True)
 
     class Meta:
+        """Meta options for CorporatePartnerCatalogLearner model."""
+
         verbose_name = "Corporate Partner Catalog Learner"
         verbose_name_plural = "Corporate Partner Catalog Learners"
         unique_together = ("catalog", "user")
 
+    # pylint: disable=no-member
     def __str__(self):
-        # pylint: disable=no-member
-        return (
-            f"<CorporatePartnerCatalogLearner: {self.user.username} in {self.catalog.name}>"
-        )
+        """Return a string representation of the CorporatePartnerCatalogLearner instance."""
+        return f"<CorporatePartnerCatalogLearner: {self.user.username} in {self.catalog.name}>"
