--- corporate_partner_access/models/__init__.py
@@ -1,9 +1,0 @@
-"""Models for corporate partner access."""
-
-from .catalogs import CorporatePartnerCatalog, FlexibleCatalogModel
-from .partner_access import (
-    CorporatePartner,
-    CorporatePartnerCatalogCourse,
-    CorporatePartnerCatalogEmailRegex,
-    CorporatePartnerCatalogLearner,
-)
