+++ flex_catalog/__init__.py
@@ -1,0 +1,1 @@
+"""Flexible catalog app - reusable catalog functionality."""
