+++ tox.ini
@@ -54,9 +54,9 @@ deps =
     -r{toxinidir}/requirements/quality.txt
 commands =
     touch tests/__init__.py
-    pylint corporate_partner_access tests test_utils manage.py setup.py
+    pylint corporate_partner_access flex_catalog tests test_utils manage.py setup.py
     rm tests/__init__.py
-    pycodestyle corporate_partner_access tests manage.py setup.py
-    pydocstyle corporate_partner_access tests manage.py setup.py
-    isort --check-only --diff tests test_utils corporate_partner_access manage.py setup.py test_settings.py
+    pycodestyle corporate_partner_access flex_catalog tests manage.py setup.py
+    pydocstyle corporate_partner_access flex_catalog tests manage.py setup.py
+    isort --check-only --diff tests test_utils corporate_partner_access flex_catalog manage.py setup.py test_settings.py
     make selfcheck
